requirements.txt
fastapi
uvicorn
python-dotenv
httpx
pydantic
cerebras-cloud-sdk
